+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - ```2019/09/30``` 修复福利界面图片无法加载的bug，修复多图片recyclerview空白处无法点击跳转的bug，新增图片预览功能，使用框架photoview，新增详情页跳转。
 - ```2019/10/09``` 添加navigationview侧滑菜单，主界面改为fragment的切换。
 - ```2019/10/14``` BaseActivity，BaseFragment 增加initOptionsCompat方法用于添加转场动画，重写startActivity,判断转场动画参数optionsCompat:ActivityOptionsCompat是否为null，如果不为null则增加，BaseActivity增加addTransitionName方法用于设置被跳转activity的转场view。
+- ```2019/10/16``` BaseActivity新增initTransitionViews方法用于初始化转场view，修改photoview的glide加载增加图片大小显示，防止有些图片被异常放大。